requirements.txt
streamlit>=1.36.0
psycopg2-binary>=2.9.9
bcrypt>=4.1.3
pandas>=2.2.0
